+++ Changelog.rst
@@ -1,6 +1,22 @@
 Change Log
 ============
 
+1.10.0
++++++++
+
+Changes
+--------
+
+* Added ``return_list`` optional argument to ``run_command`` to return list of ``HostOutput`` objects as output rather than dictionary - defaults to ``False``. List output will become default starting from ``2.0.0``.
+* Updated native clients for new version of ``ssh2-python``.
+* Manylinux 2010 wheels.
+
+Fixes
+------
+
+* Sockets would not be closed on client going out of scope - #175
+* Calling ``join()`` would reset encoding set on ``run_command`` - #159
+
 
 1.9.1
 ++++++